check if price is correct after save order

+++ commerce/src/main/java/com/myApp/order/dto/OrderDto.java
@@ -1,7 +1,7 @@
 package com.myApp.order.dto;
 
-import com.myApp.product_order.model.Product_Order;
 import com.myApp.model.UserProfile;
+import com.myApp.product_order.model.Product_Order;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -11,7 +11,8 @@ import javax.validation.constraints.Positive;
 import java.util.ArrayList;
 import java.util.List;
 
-@Getter @Setter
+@Getter
+@Setter
 public class OrderDto {
 
     @NotNull
@@ -24,7 +25,7 @@ public class OrderDto {
     private UserProfile profile;
 
     @Positive
-    private float totalPrice;
+    private double totalPrice;
 
     @Positive
     private long totalProducts;
